some changes

+++ SpringHibernateCrud/src/main/java/org/dao/IdiotDao.java
@@ -4,8 +4,11 @@ import java.io.Serializable;
 import java.util.List;
 import javax.transaction.Transactional;
 import org.hibernate.Criteria;
+import org.hibernate.Query;
+import org.hibernate.SQLQuery;
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
+import org.hibernate.criterion.DetachedCriteria;
 import org.pojo.Idiot;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
@@ -37,12 +40,18 @@ public class IdiotDao
 		Session session = sessionFactory.getCurrentSession();
 		Criteria criteria = session.createCriteria(Idiot.class);
 		return (List<Idiot>)criteria.list();
+		
 	}
 	
 	public Idiot getIdiotById(String email)
 	{
 		Session session = sessionFactory.getCurrentSession();
-		Idiot idiot = (Idiot)session.get(Idiot.class, email);
+		/*Idiot idiot  = (Idiot) session.get(Idiot.class, email);*/
+		Query q = session.createQuery("from Idiot where email = :b");
+		q.setParameter("b", email);
+		Idiot idiot = (Idiot) q.uniqueResult();
+ 		
+		System.out.println(idiot);
 		return idiot;
 	}
 
@@ -55,9 +64,8 @@ public class IdiotDao
 	public void deletetByEmail(String email)
 	{
 		Session session = sessionFactory.getCurrentSession();
-		Idiot idiot = (Idiot)session.get(Idiot.class, email);
+		Idiot idiot = (Idiot) session.get(Idiot.class, email);
+		if (idiot!=null)
 		session.delete(idiot);
 	}
-	
-	
 }
